StaffScan_Portal-main/requirement.txt
opencv-python
django
face-recognition
numpy
mysql-connector-python
opencv-python
face-recognition
Pillow
